summary/requirements.txt
torch
transformers
kobert-transformers
kss
numpy
scikit-learn
tqdm
python-mecab-ko
pymongo
python-dotenv
